--- pyproject.toml
@@ -1,24 +1,0 @@
-[tool.poetry]
-name = "plastic-detection-service"
-version = "0.1.0"
-description = ""
-authors = ["emely3h <emely3h@gmail.com>"]
-readme = "README.md"
-
-[tool.poetry.dependencies]
-python = "^3.10"
-sentinelhub = "^3.9.1"
-matplotlib = "^3.7.2"
-marinedebrisdetector = "^0.0.2"
-eo-learn = "^1.5.0"
-
-
-[tool.poetry.group.dev.dependencies]
-ruff = "^0.0.287"
-black = "^23.7.0"
-isort = "^5.12.0"
-
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
