+++ requirements.txt
@@ -11,7 +11,6 @@ boto3
 scikit-learn
 
 pytest
-
 black
 isort
 ruff
